fixed comments and added word_select to a new file.
